FInalizado ABM Rol
Queda resolver el cascadeo de los telefonos, con una logica similar a la
implementada en funciones x rol

+++ FrbaCommerce/FrbaCommerce/DAO/DaoTipoDocumento.cs
@@ -11,7 +11,7 @@ namespace FrbaCommerce.DAO
 
         static List<TipoDocumento> tiposDocumento = null;
 
-        public static TipoDocumento getTipoDocumento(Int32 id_TipoDocumento)
+        public static TipoDocumento getTipoDocumento(Decimal id_TipoDocumento)
         {
             if (tiposDocumento == null)
             {
